48 share new assignments between everybody in same course

+++ apps/api/routes/assignments.ts
@@ -1,4 +1,5 @@
 import { assignmentDBInstance, assignmentData } from "../databaseManagers/AssignmentDB";
+import { courseDBInstance } from "../databaseManagers/CourseDB";
 
 const express = require('express')
 const router = express.Router({ mergeParams: true })
@@ -19,7 +20,15 @@ router.post('/', isUserAuthorized, (req, res) => {
         userid: req.params.userid,
         weight: req.body.weight
     }
-    assignmentDBInstance.create(newAssignment)
+    courseDBInstance.userCanEditCourse(req.params.userid, req.body.courseid)
+        .then((editable) => {
+            if (editable === false) {
+                throw new Error('You do not have permission to edit this course');
+            }
+            else {
+                return assignmentDBInstance.create(newAssignment);
+            }
+        })
         .then(() => res.status(200).send(newid))
         .catch((err) => res.status(400).send(err));
 });
@@ -27,7 +36,9 @@ router.post('/', isUserAuthorized, (req, res) => {
 router.get('/', isUserAuthorized, (req, res) => {
     assignmentDBInstance.getByUserID(req.params.userid)
         .then((result) => res.status(200).send(result))
-        .catch((err) => res.status(400).send(err))
+        .catch((err) => {
+            res.status(400).send(err)
+        })
 });
 
 router.put('/:assignmentid', isUserAuthorized, (req, res) => {
@@ -35,17 +46,21 @@ router.put('/:assignmentid', isUserAuthorized, (req, res) => {
     Promise.resolve()
         .then(() => {
             if (req.body.done != null) {
-                return assignmentDBInstance.setDoneStatus(req.params.assignmentid, req.body.done)
+                return assignmentDBInstance.setDoneStatus(req.params.userid, req.params.assignmentid, req.body.done)
             }
         })
         .then(() => {
             if (req.body.grade != null) {
-                return assignmentDBInstance.setGrade(req.params.assignmentid, req.body.grade);
+                return assignmentDBInstance.setGrade(req.params.userid, req.params.assignmentid, req.body.grade);
             }
         })
-        .then(() => {
-            if (req.body.weight != null) {
+        .then(() => courseDBInstance.userCanEditCourse(req.params.userid, req.body.courseid))
+        .then((editable) => {
+            if (editable && req.body.weight != null) {
                 return assignmentDBInstance.setWeight(req.params.assignmentid, req.body.weight);
+            }
+            else {
+                throw new Error('You do not have permission to edit this assignment');
             }
         })
         .then(() => res.status(200).send('Successfully updated'))
@@ -53,7 +68,15 @@ router.put('/:assignmentid', isUserAuthorized, (req, res) => {
 });
 
 router.delete('/:assignmentid', isUserAuthorized, (req, res) => {
-    assignmentDBInstance.deleteByID(req.params.assignmentid)
+    assignmentDBInstance.userCanEditAssignment(req.params.userid, req.params.assignmentid)
+        .then((editable) => {
+            if (editable) {
+                return assignmentDBInstance.deleteByID(req.params.assignmentid);
+            }
+            else {
+                throw new Error('You do not have permission to delete this assignment');
+            }
+        })
         .then(() => res.status(200).send('Successfully deleted'))
         .catch((err) => res.status(400).send(err));
 });
